Improved API
Split checking into checking and printing

To update from old code, change

`check properties`

to

`(print << simpleCheck)`

Now that checking has been split, check takes several forms:

# Checking

* check : List Property -> Seed -> TestOutput

`check` now takes a list of properties and a random seed to feed the
process. The inclusion of the seed as a parameter make it easier to
reproduce test results and share them.

* simpleCheck : List Property -> TestOutput

`simpleCheck` is what `check` used to be. It just used an initial seed
of 1.

* continuousCheck : List Property -> Signal TestOutput

`continuousCheck` runs continuously every second and merges test
outputs. The rules for merging are simple. If the new test output is an
error, it replaces the old output, or else it doesn't. This means that
you may get an initial result of "Ok, the tests have passes" but then
later on, this result may be overwritten when an error was finally
found.

* continuousCheckEvery : Time -> List Property -> Signal TestOutput

`continuousCheckEvery` is like `continuousCheck` but you specify how
long before it retries.

* deepContinuousCheck : List Property -> Signal TestOutput

`deepContinuousCheck` is like `continuousCheck` but it just appends the
results rather than merging them. This gives you a detailed view of
what's being tested and how. (runs very second)

* deepContinuousCheckEvery : Time -> List Property -> Signal TestOutput

`deepContinuousCheckEvery` is like `deepContinuousCheck` where you can
specify how long before it retries.

# Printing

* print : TestOutput -> String

`print` converts a testOutput to a string preserving only a single
failing case per output.

* printVerbose : TestOutput -> String

`printVerbose` is like `print` but it prints all the failing test cases

* display : TestOutput -> Element

`display` displays the testOutput as an Element by calling print and
then converting it to text. This is useful to view the results in the
browser

* displayVerbose : TestOutput -> Element

`displayVerbose` is like `display` but calls `printVerbose` instead of
`print`